Pull out test helpers
